Initial moving of inputs and such so we can do prediction/input buffering/jitter buffers and so on
