+++ sabi/src/protocol/lobby.rs
@@ -1,0 +1,8 @@
+use bevy::prelude::*;
+use bevy::utils::HashMap;
+
+/// Renet Client ID -> Player Character Entity mapping
+#[derive(Debug, Default)]
+pub struct Lobby {
+    pub players: HashMap<u64, Entity>,
+}
